renamed the 'plating' package to 'templating'
