+++ keanu-project/src/main/java/io/improbable/keanu/templating/Sequence.java
@@ -1,4 +1,4 @@
-package io.improbable.keanu.plating;
+package io.improbable.keanu.templating;
 
 import org.jetbrains.annotations.NotNull;
