+++ keanu-project/src/main/java/io/improbable/keanu/templating/loop/LoopConstructionException.java
@@ -1,4 +1,4 @@
-package io.improbable.keanu.plating.loop;
+package io.improbable.keanu.templating.loop;
 
 public class LoopConstructionException extends RuntimeException {
