+++ keanu-project/src/test/java/io/improbable/keanu/e2e/foodpoisoning/FoodPoisoningTest.java
@@ -6,9 +6,9 @@ import io.improbable.keanu.KeanuRandom;
 import io.improbable.keanu.algorithms.NetworkSamples;
 import io.improbable.keanu.network.BayesianNetwork;
 import io.improbable.keanu.network.KeanuProbabilisticModel;
-import io.improbable.keanu.plating.Sequence;
-import io.improbable.keanu.plating.SequenceBuilder;
-import io.improbable.keanu.plating.SequenceItem;
+import io.improbable.keanu.templating.Sequence;
+import io.improbable.keanu.templating.SequenceBuilder;
+import io.improbable.keanu.templating.SequenceItem;
 import io.improbable.keanu.testcategory.Slow;
 import io.improbable.keanu.vertices.VertexLabel;
 import io.improbable.keanu.vertices.bool.BooleanVertex;
