+++ keanu-project/src/test/java/io/improbable/keanu/templating/SequenceBuilderTest.java
@@ -1,4 +1,4 @@
-package io.improbable.keanu.plating;
+package io.improbable.keanu.templating;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
