+++ keanu-project/src/test/java/io/improbable/keanu/templating/SequenceTest.java
@@ -1,4 +1,4 @@
-package io.improbable.keanu.plating;
+package io.improbable.keanu.templating;
 
 import org.junit.Test;
